rename IMSFeaturesVisualizer.java to IMSMobilityMzPlot
