+++ src/main/java/io/github/mzmine/modules/dataprocessing/filter_mobilitymzregionextraction/MobilityMzRegionExtractionSetupDialog.java
@@ -27,8 +27,8 @@ import io.github.mzmine.gui.chartbasics.simplechart.MultiDatasetXYZScatterPlot;
 import io.github.mzmine.gui.chartbasics.simplechart.RegionSelectionWrapper;
 import io.github.mzmine.gui.preferences.UnitFormat;
 import io.github.mzmine.main.MZmineCore;
-import io.github.mzmine.modules.visualization.imsfeaturevisualizer.CalculateDatasetsTask;
-import io.github.mzmine.modules.visualization.imsfeaturevisualizer.PlotType;
+import io.github.mzmine.modules.visualization.ims_mobilitymzplot.CalculateDatasetsTask;
+import io.github.mzmine.modules.visualization.ims_mobilitymzplot.PlotType;
 import io.github.mzmine.parameters.ParameterSet;
 import io.github.mzmine.parameters.dialogs.ParameterSetupDialogWithPreview;
 import io.github.mzmine.taskcontrol.TaskStatus;
