+++ src/main/java/io/github/mzmine/modules/visualization/ims_mobilitymzplot/CalculateDatasetsTask.java
@@ -1,4 +1,4 @@
-package io.github.mzmine.modules.visualization.imsfeaturevisualizer;
+package io.github.mzmine.modules.visualization.ims_mobilitymzplot;
 
 import com.google.common.collect.Range;
 import com.google.common.math.Quantiles;
