+++ src/main/java/io/github/mzmine/modules/visualization/ims_mobilitymzplot/IMSMobilityMzPlotParameters.java
@@ -16,8 +16,15 @@
  *  USA
  */
 
-package io.github.mzmine.modules.visualization.imsfeaturevisualizer;
+package io.github.mzmine.modules.visualization.ims_mobilitymzplot;
 
-public enum PlotType {
-  CCS, MOBILITY
+import io.github.mzmine.parameters.impl.IonMobilitySupport;
+import io.github.mzmine.parameters.impl.SimpleParameterSet;
+
+public class IMSMobilityMzPlotParameters extends SimpleParameterSet {
+
+  @Override
+  public IonMobilitySupport getIonMobilitySupport() {
+    return IonMobilitySupport.ONLY;
+  }
 }
