+++ src/main/java/io/github/mzmine/modules/visualization/ims_mobilitymzplot/PlotType.java
@@ -16,15 +16,8 @@
  *  USA
  */
 
-package io.github.mzmine.modules.visualization.imsfeaturevisualizer;
+package io.github.mzmine.modules.visualization.ims_mobilitymzplot;
 
-import io.github.mzmine.parameters.impl.IonMobilitySupport;
-import io.github.mzmine.parameters.impl.SimpleParameterSet;
-
-public class IMSFeatureVisualizerParameters extends SimpleParameterSet {
-
-  @Override
-  public IonMobilitySupport getIonMobilitySupport() {
-    return IonMobilitySupport.ONLY;
-  }
+public enum PlotType {
+  CCS, MOBILITY
 }
